api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "psycopg2-binary>=2.9.9",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    "openai[agents]>=1.3.0",
    "google-generativeai>=0.3.0",
    "httpx>=0.25.2",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "fastmcp>=2.12.0",
    "pydantic-ai-slim[duckduckgo,logfire,mcp]>=0.8.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.10",
    "sqlite-web>=0.6.4",
]
